[#47] feat: 요청내역 조회 기능 구현
