--- src/mocks/handlers/myPage/request.ts
@@ -1,7 +1,0 @@
-import { http } from "msw";
-
-/post/{postId}/request
-
-export const getRequestsHandler = http.get(`/post/bookmark`, () => {
-  return HttpResponse.json(bookmarksMockData);
-});
