+++ src/mocks/handlers/post/postHandlers.ts
@@ -6,7 +6,6 @@ import { bookmarksMockData } from '@/mocks/mockData/mypage/bookmarks';
 
 import { requestsMockData } from '@/mocks/mockData/post/requests';
 
-
 export const getAllPostsHandler = http.get('/post', () => {
   return HttpResponse.json(allPosts);
 });
@@ -54,14 +53,3 @@ export const deletePostHandler = http.delete('/post/:param', ({ params }) => {
     });
   }
 });
-
-
-export const getRequestsHandler = http.get('/post/:postId/request', ({ params }) => {
-  const { postId } = params;
-  const postIdNum = parseInt(postId as string);
-
-  if (!isNaN(postIdNum)) {
-    return HttpResponse.json(requestsMockData);
-  }
-});
-
