+++ src/types/request.ts
@@ -1,0 +1,15 @@
+export type Request = {
+  postId: number;
+  requestUser: RequestUser;
+  requestStatus: 'wait' | 'accept' | 'refuse';
+};
+
+export type RequestResponse = {
+  participationRequests: Request[];
+};
+
+export type RequestUser = {
+  id: number;
+  nickname: string;
+  manners: number;
+};
